# api/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Cloud Storage Service API
  description: |
    # Cloud Storage Service

    基于Go语言和Gin框架构建的网盘文件存储服务，提供完备的RESTful API接口。

    ## 功能特性

    - 用户认证和授权
    - 文件上传和下载
    - 文件管理和操作
    - 文件夹管理
    - 文件分享
    - 回收站功能
    - 文件版本控制

    ## 认证

    使用Bearer Token进行认证。在登录成功后，将获得的访问令牌添加到请求头中：

    ```
    Authorization: Bearer <access_token>
    ```

  version: 1.0.0
  contact:
    name: Cloud Storage Team
    email: support@cloud-storage.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: 本地开发服务器
  - url: https://api.cloud-storage.example.com/api/v1
    description: 生产服务器
...
paths:
...
  /auth/register:
    post:
      tags: [认证]
      summary: 用户注册
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequest'
      responses:
        '201':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: 请求参数错误
        '409':
          description: 用户名或邮箱已存在
        '500':
          description: 服务器内部错误
  /auth/login:
    post:
      tags: [认证]
      summary: 用户登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: 用户名或密码错误
        '403':
          description: 账户被禁用
        '500':
          description: 服务器内部错误
  /auth/logout:
    post:
      tags: [认证]
      summary: 用户登出
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 登出成功
        '401':
          description: 未认证
        '500':
          description: 服务器内部错误
  /auth/refresh:
    post:
      tags: [认证]
      summary: 刷新令牌
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: 令牌刷新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
        '401':
          description: 刷新令牌无效
        '500':
          description: 服务器内部错误
  /auth/profile:
    get:
      tags: [认证]
      summary: 获取用户资料
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: 未认证
        '500':
          description: 服务器内部错误

  # 文件相关
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    UserRegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: "john_doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        password:
          type: string
          minLength: 8
          example: "password123"
        role:
          type: string
          enum: [user, admin]
          default: user
          example: "user"
    UserLoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: "john_doe"
        password:
          type: string
          example: "password123"
    AuthResponse:
      type: object
      properties:
        message:
          type: string
          example: "login successful"
        user:
          $ref: '#/components/schemas/UserResponse'
        tokens:
          type: object
          properties:
            access_token:
              type: string
              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            refresh_token:
              type: string
              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            token_type:
              type: string
              example: "Bearer"
            expires_in:
              type: integer
              example: 3600
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    RefreshTokenResponse:
      type: object
      properties:
        message:
          type: string
          example: "token refreshed successfully"
        tokens:
          $ref: '#/components/schemas/AuthResponse/properties/tokens'

    # 用户相关
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        username:
          type: string
          example: "john_doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        role:
          type: string
          enum: [user, admin]
          example: "user"
        storage_quota:
          type: integer
          example: 10737418240
        used_storage:
          type: integer
          example: 104857600
        is_active:
          type: boolean
          example: true
        last_login_at:
          type: string
          format: date-time
          nullable: true
          example: "2023-12-31T23:59:59Z"
        created_at:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"

    # 文件相关
...